% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_transform_DB.R
\name{add_forms_transformation}
\alias{add_forms_transformation}
\title{add_forms_transformation}
\usage{
add_forms_transformation(DB, forms_tranformation, ask = T)
}
\description{
add_forms_transformation
}
