% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_transform_DB.R
\name{add_forms_transformation_to_DB}
\alias{add_forms_transformation_to_DB}
\title{add_forms_transformation_to_DB}
\usage{
add_forms_transformation_to_DB(DB, forms_tranformation, ask = T)
}
\description{
add_forms_transformation_to_DB
}
