% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_transform_DB.R
\name{add_transformation_to_DB}
\alias{add_transformation_to_DB}
\title{add_transformation_to_DB}
\usage{
add_transformation_to_DB(DB, tranformation_df, ask = T)
}
\description{
add_transformation_to_DB
}
