% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summarize_DB.R
\name{field_names_to_field_labels}
\alias{field_names_to_field_labels}
\title{field_names_to_field_labels}
\usage{
field_names_to_field_labels(field_names, DB)
}
\description{
field_names_to_field_labels
}
