% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summarize_DB.R
\name{field_names_to_form_names}
\alias{field_names_to_form_names}
\title{field_names_to_form_names}
\usage{
field_names_to_form_names(DB, field_names, only_unique = T)
}
\description{
field_names_to_form_names
}
