% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summarize_DB.R
\name{form_names_to_field_names}
\alias{form_names_to_field_names}
\title{form_names_to_field_names}
\usage{
form_names_to_field_names(form_names, DB, original_only = F)
}
\description{
form_names_to_field_names
}
