% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_setup_DB.R
\name{load_DB}
\alias{load_DB}
\title{Reads DB from the directory}
\usage{
load_DB(short_name, validate = T)
}
\arguments{
\item{short_name}{character name as a shortcut}

\item{validate}{logical for validation}
}
\value{
DB
}
\description{
Reads DB from the directory
}
