% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_transform_DB.R
\name{remove_records_from_list}
\alias{remove_records_from_list}
\title{remove_records_from_list}
\usage{
remove_records_from_list(data_list, records, id_col, silent = F)
}
\description{
remove_records_from_list
}
