% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_summarize_DB.R
\name{summarize_DB}
\alias{summarize_DB}
\title{Summarize DB}
\usage{
summarize_DB(
  DB,
  subset_name,
  drop_blanks = T,
  filter_field = NULL,
  filter_choices = NULL,
  form_names = NULL,
  field_names = NULL,
  warn_only = F,
  with_links = T,
  dir_other = file.path(DB$dir_path, "output"),
  file_name = paste0(subset_name, "_RosyREDCap"),
  separate = F
)
}
\arguments{
\item{drop_blanks}{optional logical for dropping blanks}

\item{records}{character vector of records to be summarized}
}
\description{
Summarize DB
}
