% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_app_utils.R, R/5_app_utils.R
\name{make_table1_DB}
\alias{make_table1_DB}
\title{make_table1}
\usage{
make_table1_DB(DB, df, group = "no_choice", variables, render.missing = F)

make_table1_DB(DB, df, group = "no_choice", variables, render.missing = F)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation

messages for confirmation
}
\description{
make_table1

make_table1
}
