% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_app_utils.R
\name{make_table1_DB}
\alias{make_table1_DB}
\title{make_table1}
\usage{
make_table1_DB(DB, df, group = "no_choice", variables, render.missing = F)
}
\value{
messages for confirmation
}
\description{
make_table1
}
