% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_transform_DB.R
\name{sidebar_choices}
\alias{sidebar_choices}
\title{get sidebar_choices directory}
\usage{
sidebar_choices(DB, n_threshold = 1)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\description{
get sidebar_choices directory
}
